Changed from decorator to "simple" function "changer"
A decorator is not possible for this (without ugly hacks) as the "decoration"
would require to be passed a reference to the cpu instance (for extracting the
values and storing the instruction result). So now it's a method that (after
testing) will be called during __init__() on each exec_* method to add that
value extraction & storage.